Build the basic trainer with basic testing (#3)
At this point, we have a basic training algorithm that can train the parameters of morpe, and 1 very basic unit test just to show that it runs. We do not know if the training algorithm works well or not. More testing is needed, and there is room for improvement all around.